Moved builder to separete file, removed commented code
